requirements.txt

asgiref==3.5.1
asttokens==2.0.5
attrs==21.4.0
black==22.3.0
decorator==5.1.1
django-rest-authtoken==2.1.4
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
entrypoints==0.3

PyJWT==2.4.0

social-auth-app-django==4.0.0
social-auth-core==4.3.0

tomli==2.0.1

tzdata==2022.1
anyio==3.6.1
appdirs==1.4.3
apturl==0.5.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.1
asttokens==2.0.5
attrs==21.4.0
Babel==2.10.1
backcall==0.2.0
backports.zoneinfo==0.2.1
bcrypt==3.1.7
beautifulsoup4==4.11.1
black==22.3.0
bleach==5.0.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.15.0
chardet==3.0.4
cliapp==1.20180812.1
click==8.1.3
cmdtest==0.32+git
colorama==0.4.4
command-not-found==0.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.16
debugpy==1.6.0
decorator==5.1.1
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.0
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.0.4
django-rest-authtoken==2.1.4
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
duplicity==0.8.12.0
entrypoints==0.3
executing==0.8.3
fasteners==0.14.1
fastjsonschema==2.15.3
filelock==3.0.12
fonttools==4.33.3
future==0.18.2
galternatives==1.0.6
html5lib==1.0.1
httplib2==0.14.0
idna==2.8
importlib-metadata==4.11.4
importlib-resources==5.7.1
ipykernel==6.13.0
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==7.7.0
itypes==1.2.0
jedi==0.18.1
Jinja2==3.1.2
json5==0.9.8
jsonschema==4.5.1


keyring==18.0.1
kiwisolver==1.4.3
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.1.1
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.3
mistune==0.8.4
monotonic==1.5
more-itertools==4.2.0
mypy-extensions==0.4.3
nbclassic==0.3.7
nbclient==0.5.13
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.5.5
netifaces==0.10.4
notebook==6.4.11
notebook-shim==0.1.0
numpy==1.22.4
oauthlib==3.1.0
olefile==0.46
packaging==21.3
pandas==1.4.2
pandocfilters==1.5.0
paramiko==2.6.0
parso==0.8.3
pathspec==0.9.0
pexpect==4.6.0
pickleshare==0.7.5
Pillow==7.0.0
platformdirs==2.5.2
prometheus-client==0.14.1
prompt-toolkit==3.0.29
protobuf==3.6.1
psutil==5.9.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycairo==1.16.2
pycparser==2.21
pycups==1.9.73
Pygments==2.12.0
PyGObject==3.36.0
PyJWT==2.4.0
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==3.0.9
pyRFC3339==1.1
pyrsistent==0.18.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python3-openid==3.2.0
pytz==2022.1
pyxattr==0.6.1
pyxdg==0.26
PyYAML==5.3.1
pyzmq==23.0.0
qtconsole==5.3.0
QtPy==2.1.0
reportlab==3.5.34
requests==2.22.0
requests-oauthlib==1.3.1
requests-unixsocket==0.2.0
SecretStorage==2.3.1
Send2Trash==1.8.0
simplejson==3.16.0
six==1.14.0
sniffio==1.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
soupsieve==2.3.2.post1
sqlparse==0.4.2
stack-data==0.2.0
systemd-python==234
terminado==0.15.0
testpath==0.4.4
tinycss2==1.1.1
tomli==2.0.1
tornado==6.1
traitlets==5.2.2.post1
ttystatus==0.38
typing-extensions==4.3.0
tzdata==2022.1
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.0.17
